Adds event listener to dynamically loaded incidents
