+++ client/common/js/detailsTable.js
@@ -1,3 +1,5 @@
+import {showMoreTable} from './utils'
+
 document.querySelectorAll('.list-table-toggle-btn').forEach((button) => {
 	const listId = button.getAttribute('aria-controls')
 	const list = document.getElementById(listId)
@@ -47,15 +49,5 @@ if (document.querySelector('#details-toggle-btn')) {
 }
 
 document.querySelectorAll('.incident-database-card__details-toggle').forEach((button) => {
-	button.addEventListener('click', (e) => {
-		const isVisible = Boolean(e.target.dataset.visible === 'true')
-		const newVisible = !isVisible
-		const newText = newVisible ? 'Show Less' : 'Show More'
-		e.target.innerText = newText
-		e.target.dataset.visible = newVisible.toString()
-		e.target.setAttribute('aria-expanded', newVisible.toString())
-		e.target.getAttribute('aria-controls').split(' ').forEach((id) => {
-			document.getElementById(id).dataset.visible = newVisible.toString()
-		})
-	})
+	button.addEventListener('click', showMoreTable)
 })
