+++ client/common/js/utils.js
@@ -29,3 +29,15 @@ export function isElementVisible(elm) {
 	const threshold = (rect.top + rect.bottom) * 0.75
 	return threshold < height
 }
+
+export function showMoreTable(e) {
+	const isVisible = Boolean(e.target.dataset.visible === 'true')
+	const newVisible = !isVisible
+	const newText = newVisible ? 'Show Less' : 'Show More'
+	e.target.innerText = newText
+	e.target.dataset.visible = newVisible.toString()
+	e.target.setAttribute('aria-expanded', newVisible.toString())
+	e.target.getAttribute('aria-controls').split(' ').forEach((id) => {
+		document.getElementById(id).dataset.visible = newVisible.toString()
+	})
+}
